Update default route to use Default controller and Index action
The routing pattern in the `app.MapControllerRoute` method has been
modified. The default controller has been changed from `HomePage` to
`Default`, and the default action has been changed from `HomePage` to
`Index`. This updates the default route configuration in the `Program.cs`
file, directing unspecified requests to the `Index` action of the `Default`
controller instead of the `HomePage` action of the `HomePage` controller.